FnParameterisedHolder.setParameterised() is now undoable - an additional parameter can be used to request the previous behaviour. The same parameter is also available in FnOpHolder.setOp() and FnProceduralHolder.setProcedural().
Deprecated the FnParameterisedHolder.setClassParameterClass and FnParameterisedHolder.setClassVectorParameterClasses, which were formerly the only way of editing the classes held by these parameters in a way which kept maya in sync with the cortex side changes. They weren't flexible enough for any large scale editing, and were also flawed because they were a maya specific way of editing, and we'd like to use the standard ClassParameter.setClass and ClassVectorParameter.setClasses methods on the cortex API side of things. This can now be achieved by calling the standard cortex functions mentioned inside a context manager returned by FnParameterisedHolder.classParameterModificationContext().

Because the setParameterised() method is undoable, the old create() methods of the FnOpHolder etc needed changing to prevent create() being two undoable actions as maya saw it. This meant rewriting the FnProceduralHolder.create() function to avoid using the old MEL calls, and adding an FnDagNode.createShapeWithParent() method to facilitate that.

The behaviour of the ClassParameterHandler and ClassVectorParameterHandlers also needed changing to support the changes above. The doUpdate() methods used to be implemented to set the parameter contents based on the values held in maya - that was the mechanism for restoring the classes when loading a scene. That was flawed though as the whole point of the doUpdate call is to reflect changes in the parameters into maya, rather than go the other way. Reimplemented doUpdate() to do the right thing and added new doRestore() methods specifically for the purpose of restoring things following scene load. These new methods aren't virtual right now to maintain binary compatibility, but todos items are there to change that for the next major version.